chore: moved ConfigValidator.kt to config directory
